improved seeking accuracy
git-svn-id: svn://svn.code.sf.net/p/qmmp-dev/code/trunk@827 90c681e8-e032-0410-971d-27865f9a5e38

+++ qmmp/src/ui/mainwindow.cpp
@@ -167,7 +167,7 @@ void MainWindow::replay()
     play();
 }
 
-void MainWindow::seek(int pos)
+void MainWindow::seek(qint64 pos)
 {
     if (!seeking)
         m_core->seek(pos);
